[KED-1650] Refactor graph concept
